+++ lib/manager/npm/extract/yarn.js
@@ -1,0 +1,31 @@
+const yarnLockParser = require('@yarnpkg/lockfile');
+
+module.exports = {
+  getYarnLock,
+};
+
+async function getYarnLock(filePath) {
+  const yarnLockRaw = await platform.getFile(filePath);
+  try {
+    const yarnLockParsed = yarnLockParser.parse(yarnLockRaw);
+    if (yarnLockParsed.type !== 'success') {
+      logger.info(
+        { filePath, parseType: yarnLockParsed.type },
+        'Error parsing yarn.lock - not success'
+      );
+      return null;
+    }
+    const lockFile = {};
+    for (const [entry, val] of Object.entries(yarnLockParsed.object)) {
+      logger.trace({ entry, version: val.version });
+      lockFile[entry] = val.version;
+    }
+    return lockFile;
+  } catch (err) {
+    logger.info(
+      { filePath, err, message: err.message },
+      'Warning: Exception parsing yarn.lock'
+    );
+    return {};
+  }
+}
